refactor: reorganise, rename and split components and pages
