+++ src/components/card/LoadingDefinitionCard.tsx
