--- src/components/card/index.ts
@@ -1,3 +1,0 @@
-export { default as Collection } from './Collection';
-export { default as NoFavourites } from './NoFavourites';
-export { default } from './DefinitionCard';
