+++ src/components/chapter/Chapter.tsx
@@ -4,10 +4,11 @@ import {
   useFavouriteIDs,
   useToggleFavourite,
 } from 'src/features/favouritesSlice';
-import DefinitionCard, { Collection } from 'src/components/card';
+import LoadingCollection from 'src/components/collection/LoadingCollection';
 import Error from 'src/components/error';
+import Collection from 'src/components/collection/Collection';
 import ChapterHeader from './ChapterHeader';
-import { LoadingCardList } from 'src/components/loading';
+import DefinitionCard from 'src/components/card/DefinitionCard';
 
 interface ChapterProps {
   chapter: ChapterType;
@@ -18,23 +19,21 @@ function Chapter({ chapter }: ChapterProps) {
   const favouriteIDs = useFavouriteIDs();
   const toggleFavourite = useToggleFavourite();
 
+  if (isLoading) return <LoadingCollection />;
+
   if (error) return <Error>{error.message}</Error>;
 
   return (
     <Collection header={<ChapterHeader chapter={chapter} />}>
-      {isLoading ? (
-        <LoadingCardList />
-      ) : (
-        definitions?.map((definition) => (
-          <DefinitionCard
-            key={definition.id}
-            definition={definition}
-            subheader={`${chapter.title} (page ${definition.pageNumber})`}
-            isFavourite={favouriteIDs.includes(definition.id)}
-            onBookmark={toggleFavourite}
-          />
-        ))
-      )}
+      {definitions?.map((definition) => (
+        <DefinitionCard
+          key={definition.id}
+          definition={definition}
+          subheader={`${chapter.title} (page ${definition.pageNumber})`}
+          isFavourite={favouriteIDs.includes(definition.id)}
+          onBookmark={toggleFavourite}
+        />
+      ))}
     </Collection>
   );
 }
