+++ src/components/chapter/LoadingChapters.tsx
@@ -1,0 +1,12 @@
+import LoadingCollection from 'src/components/collection/LoadingCollection';
+
+function LoadingChapters() {
+  return (
+    <>
+      <LoadingCollection />
+      <LoadingCollection />
+    </>
+  );
+}
+
+export default LoadingChapters;
