--- src/components/chapter/index.ts
@@ -1,2 +1,0 @@
-export { default } from './Chapter';
-export { default as ChapterHeader } from './ChapterHeader';
