--- src/components/loading/index.ts
@@ -1,4 +1,0 @@
-export { default as LoadingCardList } from './LoadingCardList';
-export { default as LoadingChapterHeader } from './LoadingChapterHeader';
-export { default as LoadingChapters } from './LoadingChapters';
-export { default as LoadingDefinitionCard } from './LoadingDefinitionCard';
