+++ src/pages/Quiz.tsx
@@ -1,26 +1,7 @@
-import {
-  ListItemText,
-  MenuItem,
-  MenuList,
-  Switch,
-  Tooltip,
-} from '@mui/material';
 import Box from '@mui/system/Box';
-import { useChapters } from 'src/api/chapters';
-import {
-  useSelectedChapters,
-  useToggleChapter,
-} from 'src/features/selectedChaptersSlice';
+import SelectedChaptersMenu from 'src/components/quiz/SelectedChaptersMenu';
 
 function Quiz() {
-  const { isLoading, error, data: chapters } = useChapters();
-  const selectedChapters = useSelectedChapters();
-  const toggleChapter = useToggleChapter();
-
-  if (isLoading) return <div>loading...</div>;
-
-  if (error) return <div>Error: {error.message}</div>;
-
   return (
     <Box
       height="100%"
@@ -29,31 +10,7 @@ function Quiz() {
         placeItems: 'center',
       }}
     >
-      <MenuList>
-        {chapters?.map((chapter) => {
-          return (
-            <Tooltip
-              key={chapter.id}
-              placement="right"
-              title={`${chapter.title}: ${chapter.subtitle}`}
-            >
-              <MenuItem>
-                <ListItemText>{chapter.title}</ListItemText>
-                <Switch
-                  checked={selectedChapters[chapter.id] || false}
-                  onChange={() =>
-                    toggleChapter(
-                      chapter.id,
-                      selectedChapters[chapter.id] || false,
-                    )
-                  }
-                  edge="end"
-                />
-              </MenuItem>
-            </Tooltip>
-          );
-        })}
-      </MenuList>
+      <SelectedChaptersMenu />
     </Box>
   );
 }
